test: restructure and update CarouselCreators tests
- Removed outdated tests and mocks in `src/components/carousel/__test__/CarouselCreators.test.tsx`.
- Relocated and updated snapshot to `src/components/carousel/variants/__test__/__snapshots__/CarouselCreators.test.tsx.snap`.
- Added mock files under `__mocks__` for `lens-protocol-react` and `lens-protocol-react-web`.

This improves test modularity and aligns mocks with updated dependencies.
